+++ go/app/models.go
@@ -1,0 +1,15 @@
+package app
+
+import (
+	"time"
+)
+
+type Activity struct {
+	Symbol string
+	Start  time.Time
+}
+
+type Activities struct {
+	Current Activity
+	Next    Activity
+}
